Update install script

+++ install.sh
@@ -3,18 +3,19 @@
 ## MINICONDA - need this for local installations
 #---------------------------------------------------------------------
 install_miniconda() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing miniconda......................................"
 	curl -sSL https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -o ~/miniconda.sh
 	chmod +x miniconda.sh
-	bash ~/miniconda.sh -b -p ~/.miniconda && echo "yes" | ~/.miniconda/bin/conda init
+	~/miniconda.sh -b -p ~/.miniconda && echo "yes" | ~/.miniconda/bin/conda init
 	rm -rdf miniconda.sh
-	echo export PATH="$HOME/.miniconda/bin:$PATH" >> ~/.bashrc
 	source ~/.bashrc
 }
 
 ## TMUX (TERMINAL MULTIPLEXER)
 #---------------------------------------------------------------------
 install_tmux() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing tmux..........................................."
 	conda install -y -c conda-forge tmux
 	source ~/.bashrc
@@ -23,6 +24,7 @@ install_tmux() {
 ## TMUX PLUGIN MANAGER
 #---------------------------------------------------------------------
 install_tmux_tpm() {
+	[ -d ~/.tmux ] && return
 	echo "Installing tmux-plugins/tpm......................................"
 	git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 }
@@ -30,6 +32,7 @@ install_tmux_tpm() {
 ## FUZZY FINDER
 #---------------------------------------------------------------------
 install_fzf() {
+	[ -d ~/.fzf ] && return
 	echo "Installing fzf......................................"
 	git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
 	y | ~/.fzf/install
@@ -38,6 +41,7 @@ install_fzf() {
 ## NEOVIM LATEST VERSION - STABLE
 #---------------------------------------------------------------------
 install_neovim() {
+	[ -f ~/.local/bin/nvim ] && return
 	echo "Installing neovim......................................"
 	cd ~/.local/bin
 	curl -s -JLO https://github.com/neovim/neovim/releases/download/stable/nvim-linux64.tar.gz
@@ -47,6 +51,7 @@ install_neovim() {
 }
 
 install_clang() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing clang for nvim-treesitter"
 	conda install -y -c conda-forge clang
 }
@@ -54,6 +59,7 @@ install_clang() {
 ## RUSTUP, CARGO, RUSTC
 #---------------------------------------------------------------------
 install_rust() {
+	[ -d ~/.cargo ] && return
 	echo "Installing rustup, cargo, rustc......................................"
 	curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 }
@@ -61,8 +67,9 @@ install_rust() {
 ## NODE VERSION MANAGER
 #---------------------------------------------------------------------
 install_nvm() {
+	[ -d ~/.nvm ] && return
 	echo "Installing node-version-manager......................................"
-	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.3/install.sh | bash
+	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.3/install.sh | sh
 	nvm install --lts
 	nvm install-latest-npm
 }
@@ -70,6 +77,7 @@ install_nvm() {
 ## EXA LS
 #---------------------------------------------------------------------
 install_exa() {
+	[ -f ~/.local/bin/exa ] && return
 	echo "Installing exa......................................"
 	cd ~/.local/bin
 	mkdir exa_ls && cd exa_ls
@@ -82,6 +90,7 @@ install_exa() {
 ## HTTPIE
 #---------------------------------------------------------------------
 install_httpie() {
+	[ -f ~/.local/bin/http ] && return
 	echo "Installing httpie......................................"
 	cd ~/.local/bin
 	curl -sLo http https://packages.httpie.io/binaries/linux/http-latest
@@ -91,6 +100,7 @@ install_httpie() {
 ## FD-FIND - need this for telescope live-grep & grep-string
 #---------------------------------------------------------------------
 install_fd() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing fd-find......................................"
 	conda install -y -c conda-forge fd-find
 }
@@ -98,6 +108,7 @@ install_fd() {
 ## RIPGREP - need this for telescope live-grep & grep-string
 #---------------------------------------------------------------------
 install_ripgrep() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing ripgrep......................................"
 	conda install -y -c conda-forge ripgrep
 }
@@ -105,6 +116,7 @@ install_ripgrep() {
 ## GIT-DELTA - gitdiff syntax-highlighting
 #---------------------------------------------------------------------
 install_delta() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing git-delta......................................"
 	conda install -y -c conda-forge git-delta
 }
@@ -112,17 +124,18 @@ install_delta() {
 ## BAT - just like 'cat' but with syntax-highlighting
 #---------------------------------------------------------------------
 install_bat() {
+	[ -d ~/.miniconda ] && return
 	echo "Installing bat......................................"
 	conda install -y -c conda-forge bat
 }
 
-install_miniconda
-install_tmux
-install_tmux_tpm
-install_fzf
-install_neovim
-install_exa
+install_miniconda &
+install_tmux_tpm &
+install_fzf &
+install_neovim &
+install_exa &
 install_httpie
+install_tmux
 install_fd
 install_ripgrep
 install_delta
